Fix cover and metadata display on All Tracks
