+++ pure_html/src/main/webapp/css/app.css
@@ -150,7 +150,7 @@ form{ display:grid; gap:16px; }
 }
 .track{
     display:grid;
-    grid-template-columns:40px 1fr auto auto;
+    grid-template-columns:40px 56px 1fr auto auto;
     align-items:center; gap:12px; padding:12px 14px;
     background:var(--pill); border:1px solid var(--pill-brd); border-radius:14px;
 }
@@ -165,6 +165,34 @@ form{ display:grid; gap:16px; }
 .info .a{ color:var(--muted); font-size:.95rem; white-space:nowrap; overflow:hidden; text-overflow:ellipsis; }
 .dur{ color:var(--muted); font-variant-numeric:tabular-nums; margin-right:8px; }
 .row-actions{ display:flex; gap:8px; }
+
+.cover-thumb{
+    width:56px; height:56px; border-radius:12px;
+    overflow:hidden; display:grid; place-items:center;
+    background:rgba(0,0,0,.18); border:1px solid rgba(255,255,255,.16);
+    color:var(--muted);
+}
+
+.cover-thumb img{
+    width:100%; height:100%; object-fit:cover; display:block;
+}
+
+.cover-thumb--empty{
+    font-size:1.5rem;
+}
+
+.cover-thumb__placeholder{ line-height:1; }
+
+.sr-only{
+    position:absolute;
+    width:1px;
+    height:1px;
+    padding:0;
+    margin:-1px;
+    overflow:hidden;
+    clip:rect(0,0,0,0);
+    border:0;
+}
 
 /* icon buttons (riuso con variabile dimensione) */
 .iconbtn{
@@ -197,7 +225,7 @@ form{ display:grid; gap:16px; }
 
 /* responsive stretto */
 @media (max-width:560px){
-    .track{ grid-template-columns:36px 1fr auto auto; padding:10px 12px; }
+    .track{ grid-template-columns:36px 46px 1fr auto auto; padding:10px 12px; }
     .idx{ width:36px; height:36px; }
     .dur{ display:none; }      /* nasconde durata per spazio */
     .fp-sub{ display:none; }   /* mini-player più compatto */
